btc_predict_nestquant-main/smart-career-ai-platform/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
scikit-learn==1.3.2
pdfminer.six==20231228
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.2

